rename messages csv, and move function under performance api
